Check if plymouth file is known to update-alternatives before setting
If the plymouth theme is not known to update-alternatives trying to
execute update-alternatives --set will always fail.
As plymouth kcm shows all plymouth themes in themes directory but not
all are necessarily known to update-alternatives we need to check if
they are known and if not use the update-alternatives --install instead
of --set.

Differential Revision: https://phabricator.kde.org/D5942